backend/requirements.txt
django>=6.0.4
djangorestframework>=3.15.2
django-cors-headers>=4.7.0
django-filter>=24.3
drf-yasg>=1.21.10
psycopg2-binary>=2.9.9
redis>=5.0.0
celery>=5.4.0
django-celery-beat>=2.6.0
requests>=2.31.0
channels>=4.0.0
channels-redis>=4.2.0
weasyprint>=61.2
sentry-sdk>=1.45.0
gunicorn>=21.2.0
daphne>=4.1.0
python-dotenv>=1.0.1
dj-database-url>=2.1.0
django-redis>=5.4.0
